fix timeline
