+++ pkg/infra/teams/team5/socialNetwork.go
@@ -19,7 +19,6 @@ func initTrust() AgentTrusts {
 
 type Strategy uint
 type Goodwill uint
-
 type AgentProfile struct {
 	AgentID  commons.ID
 	Trusts   AgentTrusts
@@ -41,7 +40,6 @@ const (
 	StrategyNeutral Strategy = iota
 	Chaotic         Strategy = iota
 )
-
 const (
 	Good            Goodwill = iota
 	GoodwillNeutral Goodwill = iota
@@ -59,14 +57,12 @@ type SocialNetwork struct {
 func InitSocialNetwork(ba agent.BaseAgent) SocialNetwork {
 	view := ba.View()
 	agentState := view.AgentState()
-
 	agentprofileMAP := make(map[commons.ID]AgentProfile)
 	itr := agentState.Iterator()
 	for !itr.Done() {
 		id, _, _ := itr.Next()
 		agentprofileMAP[id] = initAgentProfile(id)
 	}
-
 	return SocialNetwork{
 		AgentProfile: agentprofileMAP,
 		LawfullMin:   0.8,
@@ -75,15 +71,12 @@ func InitSocialNetwork(ba agent.BaseAgent) SocialNetwork {
 		EvilMax:      0.2,
 	}
 }
-
 func (sn *SocialNetwork) updatePersonality(agentID commons.ID, extraStrategeScore float32, extraGoodwillScore float32) {
 	agentProfile := sn.AgentProfile[agentID]
 	agentProfile.Trusts.StrategyScore += extraStrategeScore
 	agentProfile.Trusts.GoodwillScore += extraGoodwillScore
 	sn.AgentProfile[agentID] = agentProfile
-
 	sn.normaliseTrust()
-
 	if sn.AgentProfile[agentID].Trusts.StrategyScore <= sn.ChaoticMax {
 		agentProfile = AgentProfile{Strategy: Chaotic}
 	} else if sn.AgentProfile[agentID].Trusts.StrategyScore >= sn.LawfullMin {
@@ -91,7 +84,6 @@ func (sn *SocialNetwork) updatePersonality(agentID commons.ID, extraStrategeScor
 	} else {
 		agentProfile = AgentProfile{Strategy: StrategyNeutral}
 	}
-
 	if sn.AgentProfile[agentID].Trusts.GoodwillScore <= sn.EvilMax {
 		agentProfile = AgentProfile{Goodwill: Evil}
 	} else if sn.AgentProfile[agentID].Trusts.GoodwillScore >= sn.GoodMin {
@@ -99,17 +91,14 @@ func (sn *SocialNetwork) updatePersonality(agentID commons.ID, extraStrategeScor
 	} else {
 		agentProfile = AgentProfile{Goodwill: GoodwillNeutral}
 	}
-
 	sn.AgentProfile[agentID] = agentProfile
 }
-
 func (sn *SocialNetwork) normaliseTrust() {
 	var minSTG float32 = 0.5
 	var maxSTG float32 = 0.5
 	var minGW float32 = 0.5
 	var maxGW float32 = 0.5
 	var id commons.ID
-
 	for id = range sn.AgentProfile {
 		if sn.AgentProfile[id].Trusts.GoodwillScore < minGW {
 			minGW = sn.AgentProfile[id].Trusts.GoodwillScore
@@ -126,16 +115,21 @@ func (sn *SocialNetwork) normaliseTrust() {
 	}
 	distanceGW := maxGW - minGW
 	distanceSTG := maxSTG - minSTG
-	for id = range sn.AgentProfile {
-		agentProfile := sn.AgentProfile[id]
-		agentProfile.Trusts.GoodwillScore = (sn.AgentProfile[id].Trusts.GoodwillScore - minGW) / distanceGW
-		agentProfile.Trusts.StrategyScore = (sn.AgentProfile[id].Trusts.StrategyScore - minSTG) / distanceSTG
-		sn.AgentProfile[id] = agentProfile
+
+	if (distanceGW <= 1) && (distanceSTG <= 1) {
+		for id = range sn.AgentProfile {
+		}
+	} else {
+		for id = range sn.AgentProfile {
+			agentProfile := sn.AgentProfile[id]
+			agentProfile.Trusts.GoodwillScore = (sn.AgentProfile[id].Trusts.GoodwillScore - minGW) / distanceGW
+			agentProfile.Trusts.StrategyScore = (sn.AgentProfile[id].Trusts.StrategyScore - minSTG) / distanceSTG
+			sn.AgentProfile[id] = agentProfile
+		}
 	}
 }
 
 //Initialise social network
-
 // func initialiseSocialNetwork() socialNetwork {
 // 	var initialTrust uint
 // 	initialTrust = 50
